requirements.txt
# local package


# external requirements
pandas~=1.4.3
matplotlib~=3.5.3
seaborn~=0.11.2
numpy~=1.22.3
tensorflow~=2.8.0
scikit-learn~=1.1.2
#setuptools~=61.2.0
pyyaml~=6.0
sweetviz~=2.1.4
split-folders~=0.5.1
opencv-python~=4.6.0.66
imgaug~=0.4.0
tensorflow_addons~=0.18.0
tensorflow_hub~=0.12.0
dvclive~=0.10.0